+++ Dockerfile
@@ -1,0 +1,3 @@
+FROM openjdk:17
+COPY target/invincibility-points-map-0.0.1-SNAPSHOT.jar app.jar
+ENTRYPOINT ["java","-jar","/app.jar"]
