+++ src/main/java/com/invincibilitypoints/invincibilitypointsmap/controller/AuthController.java
@@ -1,4 +1,4 @@
-package com.invincibilitypoints.invincibilitypointsmap.security.controller;
+package com.invincibilitypoints.invincibilitypointsmap.controller;
 
 import com.invincibilitypoints.invincibilitypointsmap.enums.ETokenVerificationStatus;
 import com.invincibilitypoints.invincibilitypointsmap.payload.response.TokenVerificationResponse;
@@ -6,7 +6,7 @@ import com.invincibilitypoints.invincibilitypointsmap.security.payload.request.L
 import com.invincibilitypoints.invincibilitypointsmap.security.payload.request.SignupRequest;
 import com.invincibilitypoints.invincibilitypointsmap.security.payload.request.TokenRefreshRequest;
 import com.invincibilitypoints.invincibilitypointsmap.security.security.jwt.JwtUtils;
-import com.invincibilitypoints.invincibilitypointsmap.security.security.service.RefreshTokenService;
+import com.invincibilitypoints.invincibilitypointsmap.service.AuthService;
 import com.invincibilitypoints.invincibilitypointsmap.service.UserService;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
@@ -21,8 +21,7 @@ import org.springframework.web.servlet.view.RedirectView;
 @RestController
 @RequestMapping("/public")
 public class AuthController {
-    private final UserService userService;
-    private final RefreshTokenService refreshTokenService;
+    private final AuthService authService;
     private final JwtUtils jwtUtils;
     @Value("${jwt_expiration_ms}")
     int jwtExpirationMs;
@@ -30,25 +29,19 @@ public class AuthController {
     @Value("${frontend.endpoint}")
     String frontendEndpoint;
     @Autowired
-    public AuthController(UserService userService, RefreshTokenService refreshTokenService, JwtUtils jwtUtils) {
-        this.userService = userService;
-        this.refreshTokenService = refreshTokenService;
+    public AuthController(AuthService authService, JwtUtils jwtUtils) {
+        this.authService = authService;
         this.jwtUtils = jwtUtils;
     }
 
     @PostMapping("/signin")
     public ResponseEntity<?> authenticateUser(@Valid @RequestBody LoginRequest loginRequest) {
-        return userService.authenticateUserWithCredentials(loginRequest);
-    }
-
-    @PostMapping("/signup")
-    public ResponseEntity<?> registerUser(@Valid @RequestBody SignupRequest signUpRequest) {
-        return userService.registration(signUpRequest);
+        return authService.authenticateUserWithCredentials(loginRequest);
     }
 
     @GetMapping("/registrationConfirm")
     public RedirectView confirmRegistration(@RequestParam final String token) {
-        final TokenVerificationResponse result = userService.validateVerificationToken(token);
+        final TokenVerificationResponse result = authService.validateVerificationToken(token);
         if (result.eTokenVerificationStatus().equals(ETokenVerificationStatus.TOKEN_VALID)) {
             return new RedirectView( frontendEndpoint + "/successVerification" +
                     "?accessToken=" + result.jwtResponse().getAccessToken()+
@@ -59,30 +52,15 @@ public class AuthController {
     }
     @PostMapping("/refreshtoken")
     public ResponseEntity<?> refreshToken(@Valid @RequestBody TokenRefreshRequest request) {
-        return refreshTokenService.refreshToken(request);
+        return authService.refreshToken(request);
     }
-
-    @GetMapping("/passwordRecovery/sendEmail")
-    public ResponseEntity<?> sendEmailPasswordRecovery(@Valid @RequestParam String userEmail,
-                                                       HttpServletRequest request) {
-        return userService.sendEmailPasswordRecovery(userEmail, request);
-    }
-
-    @GetMapping("/passwordRecovery/checkCode")
-    public ResponseEntity<?> checkCodePasswordRecovery(@Valid @RequestParam String userEmail,
-                                                       @Valid @RequestParam String code) {
-        return userService.checkCodePasswordRecovery(userEmail, code);
-    }
-
-    @GetMapping("/passwordRecovery/update")
-    public ResponseEntity<?> updatePasswordRecovery(@Valid @RequestParam String userEmail,
-                                                    @Valid @RequestParam String code,
-                                                    @Valid @RequestParam String password) {
-        return userService.updatePasswordRecovery(userEmail, code, password);
-    }
-
     @GetMapping("/isLoggedIn")
     public ResponseEntity<?> isLoggedIn(@RequestParam final String token) {
         return jwtUtils.isLoggedIn(token);
+    }
+
+    @GetMapping("/signout")
+    public ResponseEntity<?> logoutUser() {
+        return authService.logout();
     }
 }
